feat: bindings
Data binding makes it easy to display component properties and set DOM element properties from our component to better control the view. The component can listen for and respond to events, such as a button click. And with two‑way binding, we can process user entry for an interactive experience. As we've seen, there are four basic types of binding in Angular. Here's a diagram as a memory aid. Interpolation inserts interpolated strings into the text between HTML elements or assigns element properties. Be sure to wrap the template expression in double curly braces and no quotes. Property binding sets an HTML element property to the value of a template expression. The element property must be enclosed in square brackets, and the template expression must be enclosed in quotes. Event binding listens for events from the user interface and executes a component method when the event occurs. The event name must be enclosed in parentheses, and the method to call when the event occurs must be enclosed in quotes. Two‑way binding displays a component property and updates that property when the user makes a change in an input element. Use the banana in a box syntax with the ngModel directive. The template expression must be enclosed in quotes. Here are some things to remember when using ngModel. Define ngModel within the banana in a box for two‑way binding. Be sure to add FormsModule from the Angular forms package to the imports array of an appropriate Angular module, in this case AppModule. This ensures that the ngModel directive is available to any template defined in a component associated with that module. We'll talk more about Angular modules later in this course. The data we have in our component may not be in the format we want for display. We can use a pipe in a template to transform that data to a more user‑friendly format. To use a pipe, specify the pipe character, the name of the pipe, and any pipe parameters, separated with colons. Here is an example of the currency pipe with three parameters. And here, once again, is our application architecture. In this module, we finished more of the ProductListComponent, but it could be better. Next up, we'll see several techniques for improving our component.